requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
PyPDF2==3.0.1
python-docx==1.1.0
pytest==7.4.3
httpx==0.25.2
python-multipart==0.0.6
chromadb==0.4.18
sentence-transformers==2.7.0 